requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
kombu==5.5.4
packaging==25.0
prompt-toolkit==3.0.51
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==6.1.1
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
